+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-05 13:08:42
-// Commit: fix(test): fix memory leak by cancelling test subscription
+// Updated at: 2026-05-05 15:54:39
+// Commit: chore(test): update dependencies and cleanup unused imports in test
 
 class testComponentState {
-  final String commitRef = '880';
-  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
-  final DateTime timestamp = DateTime.parse('2026-05-05T13:08:42+05:00');
+  final String commitRef = '889';
+  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
+  final DateTime timestamp = DateTime.parse('2026-05-05T15:54:39+05:00');
 }